Add dryrun target to spider a URL without storing results
